models/easyocr-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pillow>=10.0.0
pydantic>=2.0.0
numpy>=1.24.0

# EasyOCR
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
